feat(develop_liquidationPrice): add position table to future and power page
