+++ src/pages/Future/styles.less
@@ -14,6 +14,20 @@
     margin: auto;
     flex-wrap: wrap;
     justify-content: space-between;
+
+    .bottomWrapper {
+      display: flex;
+      flex-direction: column;
+      justify-content: stretch;
+  
+      .bottomTab {
+        border-top: 0.75px solid #444444;
+        border-bottom: 0.75px solid #444444;
+        padding-right: 60%;
+        margin-bottom: 0px;
+        font-size: 5px;
+      }
+    }
   }
   
   @media screen and (min-width: 1024px) {
